fixed warnings, added -Xlint, -Werror

+++ src/com/mieze/hexbattle/hex/OffsetCoord.java
@@ -14,7 +14,7 @@ public class OffsetCoord {
 
 	static public OffsetCoord qoffsetFromCube(int offset, Hex h) {
 		int col = h.q;
-		int row = h.r + (int) ((h.q + offset * (h.q & 1)) / 2);
+		int row = h.r + (h.q + offset * (h.q & 1)) / 2;
 		if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
 			throw new IllegalArgumentException("offset must be EVEN (+1) or ODD (-1)");
 		}
@@ -23,7 +23,7 @@ public class OffsetCoord {
 
 	static public Hex qoffsetToCube(int offset, OffsetCoord h) {
 		int q = h.col;
-		int r = h.row - (int) ((h.col + offset * (h.col & 1)) / 2);
+		int r = h.row - (h.col + offset * (h.col & 1)) / 2;
 		int s = -q - r;
 		if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
 			throw new IllegalArgumentException("offset must be EVEN (+1) or ODD (-1)");
@@ -32,7 +32,7 @@ public class OffsetCoord {
 	}
 
 	static public OffsetCoord roffsetFromCube(int offset, Hex h) {
-		int col = h.q + (int) ((h.r + offset * (h.r & 1)) / 2);
+		int col = h.q + (h.r + offset * (h.r & 1)) / 2;
 		int row = h.r;
 		if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
 			throw new IllegalArgumentException("offset must be EVEN (+1) or ODD (-1)");
@@ -41,7 +41,7 @@ public class OffsetCoord {
 	}
 
 	static public Hex roffsetToCube(int offset, OffsetCoord h) {
-		int q = h.col - (int) ((h.row + offset * (h.row & 1)) / 2);
+		int q = h.col - (h.row + offset * (h.row & 1)) / 2;
 		int r = h.row;
 		int s = -q - r;
 		if (offset != OffsetCoord.EVEN && offset != OffsetCoord.ODD) {
